Patch by Peter Kjellerstedt for cleaning up after broken versions of SCCS.
Some non y2k-compliant versions of SCCS will use a ":" in the decade
field of the date in the year 2000 (since the ASCII code for ':' is
one greater than that for '0').  Changed the relevant sccs_date
constructor to figure out the right value in this case.  A warning
message is issued when this happens.