Fix from Frank van Maarseveen <fvm@tasking.nl>:
A setuid cdc/rmdel will not unlock (i.e. remove the z-file) in
case of an error. Most obvious error is when the invoker is not
the creator of the delta.

Both cdc and rmdel revoke setuid permission when the real user
is not the creator of the delta. However, this should be restored
_before_ reaching the end of the 'try' scope because on leaving
it the z-file is removed indirectly by the sccs_file destructor.

Attached you will find a patch which solves this problem.

A more correct solution would be not to swap uid/euid but
to print something like:

ERROR [/usr/src/xxx/SCCS/s.main.c]:
Must be owner of file or <user> to use this command.(rc4)