Allow an empty MR list to be passed to admin.
* tests/admin/init-mr.sh: Don't expect "admin -m" to interpret
a plain "-m" as meaning an empty MR list.  Distinguishing the
absence of an argument to -m requires a traditional SCCS argument
processor, and this behaviour is incompatible with modern
implementations of SCCS (for example OpenSolaris).
* tests/admin/flags.sh: Likewise.
* src/admin.cc (main): Allow an empty list of MRs to be specified
to admin, for example "admin -fv -m' '".  OpenSolaris allows
this, and we can no longer use the previous method which was to
omit the optional argument.  Elminiate doubled space in error
message.
* docs/cssc.texi (admin): document this clearly.
* NEWS: mention this change.

Remove duplicate space in "MR number(s) must be  supplied."

* src/admin.cc (main): elminiate doubled space in error message.