Add detail message to Failure.  Create a FailureBuilder class.
* src/failure.h (Failure): add a string field which holds additional
details.  Rename message() to to_string() to avoid confusion now that
the message is not going to be identical to the result of
std::error_code::message().  Add make_failure...() factory functions
allowing the user to provide a detail message.
(FailureBuilder): new class allowing a Failure instance to be built
and have details added using the "<<" insert operator, to take
advantage of the ability to format arbitrary objects.  Use a
std::ostringstream to do the formatting.   Also add a diagnose()
method which causes the FailureBuilder instance to emit failure
message when the destructor is called (which will usually be at the
completion of the enclosing full expression, since the FailureBuilder
instance will rarely be given a name).
* src/failure_or.h (FailureOr<T>): allow construction from a
FailureBuilder instance.  Rename the message() member function to
to_string() for consistency with the change to Failure.
* src/failure.cc: Implement FailureBuilder methods, rename message()
to to_string().  Implement new factory functions.
* src/body-scanner.cc: use the new FailureBuilder factory functions to
issue error messages when failures are detected.
* src/bodyio.cc, src/file.cc, src/parser.cc: Likewise.