Change one of the sccs_file::get() methods to return Failure.
* src/body-scanner.cc (sccs_file_body_scanner::get): change return
type from bool to Failure.  Use a longish lived Failure variable to
record whether somethign went wrong.  Annotate control paths with
unreachable lines (following a call to corrupt()).
* src/body-scanner.h (sccs_file_body_scanner::get): change return
type.
* src/sf-get.cc (sccs_file::get): change return type.
* src/writesubst.cc (sccs_file::write_subst): wrap the whole function
in a lambda whose body is mostly unchanged - that is, it returns an
int.   Convert the int result into a csssc::Failure object after
calling the lambda.  This is a temporary situation which we should be
able to eliminate as we get rid of the GetFileBodyFailed error code.
* src/sf-get2.cc (sccs_file::get): use more specific failure code
CannotEditBitkeeperFile.
* src/sf-prs.cc (sccs_file::get): Likewise.
* src/sf-rmdel.cc (sccs_file::rmdel): Likewise.
* src/failure.h (errorcode): add CannotEditBitkeeperFile.
(ok): remove, and instead consistently use cssc::Failure::Ok().
* src/failure.cc (category_impl::message): add a message for
CannotEditBitkeeperFile.
* src/get.cc: Use cssc::Update to record whether something has
failed.
* src/linebuf.cc: use cssc::Failure::Ok() not cssc::ok().
* src/sccsfile.h (get): update return type.
* src/sf-delta.cc (sccs_file::add_delta): update to new return type
for get().
* src/sccsfile.cc (edit_mode_ok): rename to edit_mode_permitted and
make this return Failure, so that when this situation occurs, the
caller can simply return the ready-made Failure object.
* src/sf-csc.cc (sccs_file::cdc): call edit_mode_permitted instead of
edit_mode_ok.