Disallow initialising FailureOr<T> with a Failure::Ok().
* src/failure_or.h (FailureOr::FailureOr(const Failure&)): disallow
promoting "OK" to FailureOr<T> since this leaves us with a non-empty
FailureOr with a default-constructed T.
* src/writesubst.cc (sccs_file::emit_keyletter_expansion): don't do
that.  It happens that the value we should have passed instead was
false, which is a default-constructed bool, so there is no actual
difference in behaviour.
(sccs_file::write_subst): slightly simplify the control flow.