Put the copyright data in a separate include file.
Our previous attempt did not work because the \n" at the end of the
lines in the copyright string constant made "make update-cpoyright"
fail to recognise it as a copyright statement of interest.
* src/copyright.cc: move the copyright data to a #include directive.
Since the copyright string constant is no longer in this file,
reinstate the header comment.
* src/copyright_data.txt: the copyright statement is moved to here; we
process it to make a generated source file, copyright_data.inc.  That
file is included in copyright.cc.
* src/Makefile.am: add rules to generate copyright_data.inc and a rule
which tells Make that copyright.cc includes that file (so that we know
when to rebuild copyright.$(OBJEXT)).
* src/copyright.awk: this is the program which generates
copyright_data.inc from copyright_data.inc.tmpl and
copyright_data.txt.
* src/copyright_data.inc.tmpl: the template for copyright_data.inc.